ENH: new plot directive from matplotlib svn
This imports the plot directive:

  sphinxext/plot_directive.py

from the v0_99maint branch, version 8988, of matplotlib svn repository.

I chose the 0.99 version to remain backwards compatible, the 1.00
plot_directive.py will not play nicely with sphinx 0.99. Note: Ubuntu 10.4 ships
with 0.99.

In this version, the links of the form:

  [source code, hires.png, pdf]

Are created, so for every plot we make with this directive, the source code is
included. The was broken with our old version, see also:

  https://github.com/scipy-lectures/scipy-lecture-notes/issues#issue/3

This unfortunately creates another problem, because the :no-links: seems no
longer supported, perhaps there is a new way to do this now.

The next problem is that instead of directly reading from the pyplot directory,
the full path of the code must be specified, this is not a real problem, it just
required adaptation of the existing sources.

Also :align: center still does not work.

The other cool thing, is that all the plot output, (png,pdf) etc is put directly
into the build directory, so now 'make clean' actually cleans, and no need to do
cumbersome 'git clean -dfx ; git clean -dfX'