Move from sym.solve to sym.solveset.
This is a fork of
https://github.com/scipy-lectures/scipy-lecture-notes/pull/307 and
closes #308

This also fixes a number of outputs that were not being doctested (the
Jupyter `In` `Out` prompt was being used which is not recognised by the
python doctest module).

Note that I did not change `sym.solve` for the case of the system of
equations (the alternatives are not yet implemented solveset and I feel
make things more complicated than they need to be). FWIW in my own
teaching I do not consider systems of linear equations using any of the
solvers and instead show students how to use Numpy (or indeed Sympy) to
do this using matrices.