Use empty list instead of empty tuple to deactivate ticks
This is the canonical expression recommended by matplotlib. See the last
example at
https://matplotlib.org/api/_as_gen/matplotlib.pyplot.xticks.html

While technically any zero-length iterable works, it's more clear to
use a list. Firstly, ticks are conceptually arbitrary-length iterables,
which is better expressed with a list type. Secondly, the empty square
bracket list expression is more widely known compared to the empty
parentheses expression for a tuple.